+++ backend/src/Constants.ts
@@ -2,3 +2,11 @@ export const MAXIMUM_LENGTH_OF_USER_NAME = 20;
 export const MINIMUM_LENGTH_OF_USER_NAME = 5;
 export const MAXIMUM_LENGTH_OF_USER_PASSWORD = 40;
 export const MINIMUM_LENGTH_OF_USER_PASSWORD = 5;
+
+export const DefaultLanes = [
+  { name: 'Not Qualified', inForecat: true },
+  { name: 'Qualified', inForecat: true },
+  { name: 'Comitted', inForecat: true },
+  { name: 'Closed Won', color: '#00b359', inForecat: false },
+  { name: 'Closed Lost', color: '#e30544', inForecat: false },
+];