+++ frontend/src/components/Board.tsx
@@ -1,58 +1,14 @@
-import { useDragDropManager } from 'react-dnd';
 import { Lane } from './Lane';
-import { Trash } from './Trash';
 import { useSelector } from 'react-redux';
-import { useState, useEffect, useContext } from 'react';
-import { selectCards, store } from '../store/Store';
-import { ActionType } from '../actions/Actions';
-import { Lane as LaneInterface, LaneKey } from '../Constants';
-import { RequestHelperContext } from '../context/RequestHelperContextProvider';
+import { selectCards } from '../store/Store';
+import { Lane as LaneInterface } from '../interfaces/Lane';
 
 export interface BoardProps {
   lanes: LaneInterface[];
 }
 
 export const Board = ({ lanes }: BoardProps) => {
-  const { client } = useContext(RequestHelperContext);
-
   const cards = useSelector(selectCards);
-  const dragDropManager = useDragDropManager();
-  const [isDragging, setIsDragging] = useState(false);
-
-  useEffect(() => {
-    dragDropManager.getMonitor().subscribeToStateChange(handleMonitorChange);
-  }, []);
-
-  const handleMonitorChange = () => {
-    setIsDragging(dragDropManager.getMonitor().isDragging());
-  };
-
-  // TODO should be a redux action
-  const moveCard = async (id: string, key: LaneKey) => {
-    console.log(`move card ${id} to lane ${key}`);
-
-    const card = cards.find((card) => card.id === id);
-
-    if (card) {
-      if (key === 'trash') {
-        await client!.deleteCard(card!.id);
-
-        store.dispatch({
-          type: ActionType.CARD_DELETE,
-          payload: card!.id,
-        });
-      } else {
-        card!.lane = key;
-
-        await client!.updateCard(card);
-
-        store.dispatch({
-          type: ActionType.CARD_UPDATE,
-          payload: card,
-        });
-      }
-    }
-  };
 
   return (
     <>
@@ -64,7 +20,6 @@ export const Board = ({ lanes }: BoardProps) => {
           return (
             <Lane
               key={lane.key}
-              moveCard={moveCard}
               lane={lane}
               cards={list}
               numberOfLanes={
@@ -73,7 +28,6 @@ export const Board = ({ lanes }: BoardProps) => {
             />
           );
         })}
-      {isDragging && <Trash isDragging={isDragging} />}
     </>
   );
 };