+++ frontend/src/components/Lane.tsx
@@ -1,33 +1,21 @@
 import { useEffect, useState } from 'react';
-import { Card } from './card/Card';
-import { ItemTypes } from './ItemTypes';
-import { useDrop } from 'react-dnd';
-import { Lane as LaneInterface } from '../Constants';
+import { Card } from './Card';
+import { Lane as LaneInterface } from '../interfaces/Lane';
 import { Card as CardInterface } from '../interfaces/Card';
-import { selectCurrency } from '../store/Store';
+import { selectCurrency, store } from '../store/Store';
 import { useSelector } from 'react-redux';
+import { Droppable } from 'react-beautiful-dnd';
+import { ActionType } from '../actions/Actions';
+import { Currency } from './Currency';
 
 export interface LaneProps {
-  moveCard: any;
   lane: LaneInterface;
   cards: CardInterface[];
   numberOfLanes: number;
 }
 
-export const Lane = ({ moveCard, lane, cards, numberOfLanes }: LaneProps) => {
-  const currency = useSelector(selectCurrency);
+export const Lane = ({ lane, cards, numberOfLanes }: LaneProps) => {
   const [amount, setAmount] = useState(0);
-
-  const [{ isOver }, drop] = useDrop(
-    () => ({
-      accept: ItemTypes.CARD,
-      drop: () => ({ name: lane.key }),
-      collect: (monitor) => ({
-        isOver: !!monitor.isOver(),
-      }),
-    }),
-    []
-  );
 
   useEffect(() => {
     setAmount(
@@ -37,10 +25,18 @@ export const Lane = ({ moveCard, lane, cards, numberOfLanes }: LaneProps) => {
     );
   }, [cards]);
 
+  const showLaneDetail = (id?: string) => {
+    store.dispatch({
+      type: ActionType.USER_INTERFACE_STATE,
+      payload: { state: 'lane-detail', id: id },
+    });
+  };
+
   return (
     <div className="lane" style={{ width: `${100 / numberOfLanes}%` }}>
       <div
         className="title"
+        onClick={() => showLaneDetail(lane.key)}
         style={{
           backgroundColor: lane.color ?? '#e6e6e6',
           color: lane.color ? 'white' : 'grey',
@@ -58,28 +54,32 @@ export const Lane = ({ moveCard, lane, cards, numberOfLanes }: LaneProps) => {
       >
         {cards.length} Deal{cards.length > 1 ? 's' : ''} -{' '}
         <b>
-          {amount.toLocaleString('en-US', {
-            style: 'currency',
-            currency: currency ?? 'USD',
-          })}
+          <Currency value={amount} />
         </b>
       </div>
-
-      <div
-        className="canvas"
-        ref={drop}
-        style={{
-          backgroundColor: isOver ? '#F6F6F6' : 'white',
+      <Droppable droppableId={lane.key}>
+        {(provided, snaphot) => {
+          return (
+            <div
+              ref={provided.innerRef}
+              {...provided.droppableProps}
+              className="canvas"
+              style={{
+                backgroundColor: snaphot.isDraggingOver ? '#F6F6F6' : 'white',
+              }}
+            >
+              {cards
+                .filter((card) => card.lane === lane.key)
+                .map((card, index) => {
+                  return (
+                    <Card index={index} key={card.id} card={card} lane={lane} />
+                  );
+                })}
+              {provided.placeholder}
+            </div>
+          );
         }}
-      >
-        {cards
-          .filter((card) => card.lane === lane.key)
-          .map((card) => {
-            return (
-              <Card key={card.id} card={card} moveCard={moveCard} lane={lane} />
-            );
-          })}
-      </div>
+      </Droppable>
     </div>
   );
 };