+++ frontend/src/interfaces/CurrencyCode.ts
@@ -1,6 +1,0 @@
-// ISO 4217 currency code
-export enum CurrencyCode {
-  USD = 'USD',
-  EUR = 'EUR',
-  SEK = 'SEK',
-}