+++ frontend/src/store/ApplicationStore.ts
@@ -1,9 +1,11 @@
 import { CurrencyCode } from '../interfaces/Account';
 import { BrowserState } from '../interfaces/BrowserState';
 import { Card } from '../interfaces/Card';
+import { Lane } from '../interfaces/Lane';
 
 export interface ApplicationStore {
   cards: Card[];
+  lanes: Lane[];
   session: {
     token: string | undefined;
     alerts: number;
@@ -21,7 +23,7 @@ export interface ApplicationStore {
     isPageLoaded: boolean;
   };
   ui: {
-    state: 'default' | 'card-detail';
-    id: undefined | Card['id'];
+    state: 'default' | 'card-detail' | 'lane-detail';
+    id: undefined | Card['id'] | Lane['key'];
   };
 }