+++ frontend/src/store/Store.ts
@@ -17,6 +17,9 @@ export const selectIsPageLoaded = (store: RootState) =>
 export const selectBrowserState = (store: RootState) => store.browser.state;
 export const selectToken = (store: RootState) => store.session.token;
 export const selectCards = (store: RootState) => store.cards;
+export const selectLanes = (store: RootState) => store.lanes;
+export const selectLane = (store: ApplicationStore, key: string) =>
+  store.lanes.find((lane) => lane.key === key);
 export const selectName = (store: RootState) => store.session.user.name;
 export const selectCurrency = (store: RootState) =>
   store.session.account.currency;