+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## 2023-02-07
+
+- wrapped all forecast values in a `Currency` component
+- added Lanes to Redux store and backend
+- switched drag-and-drop functionality to react-beautiful-dnd
+
 ## 2023-02-06
 
 - Added currency support